test: refactor directory structure
This commit refactors the `test` directory to mirror `lib`. The tests
are also distinguished between unit and integration tests. The unit
tests for each module in `lib` can be found in a directory of the same
name in `test`. The `integration` directory stores all integration
tests. Furthermore, the `-test.js` suffix has been replaced with
`-unit.js` and `-integration.js` to allow npm scripts to target test
types in isolation.

Co-Authored-By: Boyma Fahnbulleh <boymanjor@protonmail.com>