rpc: new wallet rpc getaddressinfo
Adds a new wallet RPC method getaddressinfo that
corresponds to changes with bitcoind. Returns values:

- address
- ismine
- iswatchonly
- ischange
- isspendable
- isscript
- witness_version
- witness_program

The ismine property previously existed on the node RPC
method validateaddress which was always hardcoded
to true.

Commit 62b4de584cdabd4488d0583cd4c2b2f3e75d4a60
updates `validateaddress` to remove the `ismine`
property because that property can now be found
on the `getaddressinfo` RPC.

The `isspendable` property refers to an address
will null data, which is version 31.

See PR in bcoin: https://github.com/bcoin-org/bcoin/pull/731
See bitcoind docs:
https://bitcoincore.org/en/doc/0.17.0/rpc/wallet/getaddressinfo/