node/rpc: render peer fullname in getpeerinfo
Use the peer method fullname to render the
full address returned in the peer json property
addr. This matches the hosts.json file, where
addr is pubkey@ip:port. This is useful information
for building automation and helpful when sharing
peers with other people, so that the hosts.json
file doesn't need to be inspected. Also use the
fullname method to render the addrlocal property.