node rpc: add identitykey to getpeerinfo
Add the peer's identity key to each peer object
that is returned by the node RPC getpeerinfo.
This is useful information to build automation
on top of and also helpful for sharing with other
people instead of having to look at the hosts.json
file to share the node's peers.