net/peer: add fullname and localname
Add fullname and localname methods to the
Peer class. fullname wraps the peer's
address.fullname and localname returns
the local address's fullname. Since
the local address should share the same
key as the address, there is no need to
hold both in memory. localname renders
the the full hostname using the address's
key along with the local hostname and port.