readme: udpate mining instructions
Users are confused because the Proof of Work algorithm has
changed and hs-miner is no longer the best way to mine.
It was observed during testnet3 that hashrate for the hsd
miner was far greater than the hashrate for hs-miner. There
is also a bug in hs-miner regarding an incorrect header size.
This PR updates the README.md to only mention the hsd CPU miner.
Community members are working on a GPU miner that is not
out of beta yet.

See:
https://github.com/handshake-org/hs-miner/issues/12
https://handshake.community/t/error-bad-header-size-when-starting-miner/66