wallet/http: allow creation of unsigned tx
Add a `sign` parameter that defaults to true.
The created transaction will only be signed if
`sign` is true. This makes it much easier to
create transactions for wallets where the keys
are not stored on the server. The client can pass
`sign: false` and the request will return an
unsigned transaction that can be signed using
a hardware device.