exports: export Covenant along with other primitives
This commit simply adds the Covenant to the list of exports. It's a
primtive that I believe is used quite a bit by external developers, so
it would be very useful to have it exported easily rather than having
the developers manually point to the path, especially considering that
every other primitive that makes up a block -> transaction is exported.