test: contractstate
Adds tests for the ContractState class that is used
in the mempool to track the state of the chain along
with transactions in the mempool.

An instance of the ContractState class has a series
of maps for different types of name related outputs.
These tests create txs containing name related outputs
and assert that the ContractState indexes them
correctly based on their covenant type.