wallet: create open lock
There are many different types of transactions
regarding the auction system. Each covenant
type has its own explicit code path for creation.
Currently, only the send functionality has a
lock around it which means that there is no lock
being used for code that wants to create the
transaction but not sign and/or broadcast it.
This pull request adds a new method for each
create transaction type method that wraps the
actual method with a lock.

Since there is no enforced concept of private
methods in Javascript, this makes each of the
create transaction methods safe for external
consumption. These functions were only being
called from within the wallet before, but in
an attempt to better support watch only wallets,
the create transaction methods would like to
be consumed from outside of the wallet itself.

An example:

// original createBid functionality here
_createBid() {}

createBid() {
  // create lock
  // call _createBid
  // unlock
}

The above pattern was followed for each of the
create methods. The send methods call the create
methods and there is already a lock around the
send methods (they follow a similar pattern as
above), they were updated to directly use the
methods with the underscores.