net: jsdoc fixes for constants
This commit updates the documenation on two constants in the net common
file. The first is the required services which are now only NETWORK and
not NETWORK and SEGWIT. The second is an update to the maximum message
size. The max message size was updated from 4mb to 8mb in commit: 4c5668994162053be6c02d3af5c7959b6a572b47