node/http: return interval height and blocks until
To prevent bugs on the client side, return
both the number of blocks until the next
tree committment and return the height of
the next tree committment. Application builders
may need either value and while it is possible
to compute the other with one, its better
to just return both.