wallet/http: create renewal
Implement the HTTP endpoint
`POST /wallet/:id/renewal`. This creates
a renewal transaction, which is required
to be sent every so often (once per year on
mainnet) to ensure that names are not
locked up in a lost key and to pay for
state rent. If no renewal tx is sent,
the name goes back up for auction.