script: sighash SINGLEREVERSE
Implement a new sighash type SINGLEREVERSE. This
sighash type is similar to SINGLE, but is different
because instead of committing to the same output
index as the input's index being signed, it commits
to the index at (outputs.length-1) - index.

If the first input was signed with SINGLEREVERSE,
then it would only commit to the final output.
This is useful because covenants are "linked", so
that the next covenant in the series must be at the
same index as the input. This new sighash type allows
a user to commit to the covenant input but not the
output for that covenant.

This is a building block for atomic name swaps on chain.