blockchain: emit tree commit event from chain
Emit an event named 'tree commit' from the chain
on the 'chain' topic. It emits three items,
the newly comitted tree root (which is not included
in a block until the next block) along with the
current block and the chain entry.

This happens in the `setBestChain` method on the chain.
The computation to determine the event is:

```
if ((entry.height % this.network.names.treeInterval) === 0)
```

The call stack for a block being added to the chain is as follows:

- `Chain.add` - [source](https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L1885)
- `Chain._add` - [source](https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L1904)
- `Chain.connect` - [source](https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L1979)
- `Chain.setBestChain` - [source](https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L1624)

The logic inside of `Chain.connect` determines if `Chain.
setBestChain` will be called or if `Chain.saveAlternate` is called.

The relevant codepath:

https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L2008-L2021

If the incoming block's chainwork is less than or equal to
the current tips chainwork, then `Chain.saveAlternate` is
called. The only way that `Chain.setBestChain` is called
is if the incoming block has greater chain work. This means
that the jsdoc is __incorrect__ for `setBestChain` regarding
the line `"This is called on every valid block that comes in."`.
It is only called for blocks that extend the current chain
and not for alternative blocks.

https://github.com/handshake-org/hsd/blob/e1c0281dd9e9bbbd23f5cd832c82ebe3808e9b4f/lib/blockchain/chain.js#L1611-L1625

You can see here that the jsdoc is correct regarding `setBestChain` as a private function. The
only place that it is invoked is withing `lib/blockchain/chain.js` inside of `Chain.connect`.

```bash
$ basename $PWD
hsd
$ grep -rn setBestChain lib
lib/blockchain/chain.js:1464:    // That will be done outside in setBestChain.
lib/blockchain/chain.js:1624:  async setBestChain(entry, block, prev, flags) {
lib/blockchain/chain.js:2020:      await this.setBestChain(entry, block, prev, flags);
```

After doing this review, I believe that it is safe to include
the event emitting exactly where it is currently in this pull
request. `Chain.setBestChain` is not called for every block,
just blocks extending the best chain.