blockchain: emit tree commit event on reconnect
This adds the `"tree commit"` event to `Chain.reconnect`
which is part of the codepath for reorgs. This way the
event is triggered when reorgs occur. Blocks only ever go
through `setBestChain` once, every other time that they
are reorged out and then reincluded in the chain, it will
be through the `reconnect` method.