net: remove discoverExternal() from the pool
This commit removes the discover external check in the pool. Currently
we already store all local addresses that are sent to us from peers in
the hostlist, so there is no reason to rely on an external service to
check for our local/external address. This was removed from bitcoin in
commit 845c86d.