sighash: SINGLEPRESIGNFINALIZE
Implement new sighash type SINGLEPRESIGNFINALIZE.
This sighash type is the same as sighash SINGLE
but will not commit to the item at index 6 in the
covenant items list only if the covenant type is
FINALIZE. The item at index 6 is the blockhash.
This blockhash must come one transfer lockup period's
amount of blocks ahead of the block that the transfer
was included in the blockchain.

It adds risk regarding the two transaction name trasfer
scheme, but the risk is bounded. Without this sighash
type, the attacker must have access to the key directly
when they send the TRANSFER transaction and then must have
access to the key again when they send the FINALIZE transaction,
because they must commit to a blockhash from the future.
If they cannot have access to the key twice (can't copy it)
then they will not be able to steal the name. It is possible
for the attacker to attempt to brute force the FINALIZE
transaction by signing many different onces with many different
blockhashes. This attack becomes easier as the difficulty goes up.

If the attacker is able to copy the key, then they will be able
to send the second transaction in the future after observing the
blockhash themselves. The nameholder must observe the chain
for any TRANSFER transactions on names that they did not authorize
and then send an UPDATE transaction to cancel the name transfer.
If the attacker has a copy of the key, they can attempt to
TRANSFER the name again, so the last resort is to REVOKE the
name.

The final building block for non-interactive name swaps.
I believe this is very important for the security of the
chain over the long run because transactions must be able
to support the security of the network after the miner reward
runs out. People will be able to list their names on a website
with 2 presigned transactions, the TRANSFER and FINALIZE. The
name must be locked in a script with a policy that only allows
for spends to FINALIZE using OP_TYPE. This is risky because
a stolen key while locked in this script will result in the
name being stolen, but in return the name is able to be traded
on chain in an atomic way without giving up any personally
identifiable details.