net-pool: Check outbound peers for same group before connecting
This commit checks the outbound peers to see which group their network
addresses lives in before connecting. By doing this, we prevent a node
from connecting to all outbound addresses in a very close network group
bunch. This helps prevent someone spinning up multiple nodes on a
similar network e.g. AWS, GCP and using those to fill other nodes'
outbounds.

This commit just adds 2 operations in both addOutbound and addLoader. It
first checks if the new address pull from hostlist belongs in the same
group as an already connected peer. If it does not, it precedes as
usual, and if the connection succeeds then it adds that new outbound
group to the list of groups currently connected to.

Co-authored-by: Nodari Chkuaselidze <nodar.chkuaselidze@gmail.com>