test: add mining test for new proof of work
This commit adds multiple tests to test the new proof of work function.
Mainly, mining a block is tested, as well as mining a block with a mask
that mimics what they would receive from a pool.

We also added in an extra parameter to the mine function to allow for a
nonce to be passed in instead of starting with 0. This should have zero
side effects as if it is not included, we default to the previous value
(0). We also add a funciton named mineRandomNonce that uses a random
nonce value to start iterating from - simulating a pool-like
environment.

Co-Authored-By: Mark Tyneway <mark@purse.io>
Co-Authored-By: Sean Kilgarriff <seanpkilgarriff@gmail.com>