node/http: add mempool details param
GET /mempool now accepts a query param
`details=bool`. When set to true, it
will return the TX JSON instead of just
the hashes. This is useful for block
explorers and also anybody who is interested
in more details about what exactly is in the
mempool at any given time.