primitives/address: add toScript method
This converts the Address into a Script.
This same functionality occurs in Script.exeucte,
as the Address.hash needs to be templated into
a Bitcoin Script to run through the interpreter.
In reality, Address.hash does not need to be
a hash. To know what the script will be,
the size of Address.hash must be observed. In
the case of 20 bytes, it is known to be a pay
to pubkey hash, which can be created with
Script.fromPubkeyHash. If the Address.hash
is 32 bytes, the witness must be present
to build the script correctly. In the case
of Address.version === 31, it is an OP_RETURN
script. Unknown Address types will return `null`.