dns/records: add cosmoshub support
Unknown currencies are saved on chain in
ascii format as currency:address.
Adding 'cosmoshub' as a known currency
will save space for saving these types of
addresses on chain, as the hrp will be encoded
in a field instead of saving the ascii to chain.
This is particularly helpful for cosmoshub
addresses, as the hrp is relatively large
compared to other chains.

The sizes below are in bytes

hrp           | size before  | size after |
--------------|--------------|------------|
cosmos        | 59           | 24
cosmosvalcons | 66           | 24
cosmosvaloper | 66           | 24

Support for all 6 hrps is included
- cosmos
- cosmospub
- cosmosvalcons
- cosmosvalconspub
- cosmosvaloper
- cosmosvaloperpub

This reduces the size of the cosmoshub addresses
on chain by more 36-40%.