dns/record: fix ssh record object property
The Handshake `SSH` record type was defined
to have a `keyType`, but the `bns` based
`SSHFPRecord` was defined with a `digestType`.
This commit changes the `keyType` property
to being called `digestType`, as the data
was not being properly assigned in `hsd/lib/resource`
when the `Resource` was being rendered into a
`sshfp` DNS request.

See this issue for more details:
https://github.com/handshake-org/hsd/issues/285