wallet: mitigate watch only wallet bug for getKey
This PR mitigates a bug which potentially stores an incorrect
account index in the db of watch only wallets. This is only a
patch of a specific API endpoint. It is necessary for recent
work on ledger-app-hns. Specifically, we have implemented
on-device output verification and need to be able to retrieve
a reliable path to verify and skip change outputs.

Future work involves fixing the source of this issue. For more
details see: https://github.com/bcoin-org/bcoin/issues/698.