Cleaning up Dockerfile with Best Practices
Removed the `apt update` command as the best practice way to use a 3rd party Docker image as base layer is to keep the package versions that come with it.  Also might consider pinning the versions of each package added in line 14 to known working versions in the case there's any breaking changes (unlikely with those specific packages so I left it alone.