test: abstract some mining heights
It is much more useful to mine based
on the auction lengths instead of using
magic numbers. This allows the parameters
to be tweaked and then ran through the
test suite because it is possible to choose
parameters that do not allow the auctions
to work at all. The `Network.get` call
at the top of the file can be altered
to select a different network, such as
simnet.