miner: temporarily disable softfork voting by default.
Several mainnet miners have been mistakenly voting on the RSA hardening
soft-fork. This is a result of pool software creating the template internally,
a la bstratum (i.e. not via getwork or getblocktemplate).

Until miner code either accounts for this or switches to GBT completely, we
should take measures to avoid this.