wallet: new rpc getblindvalue
An imported wallet may "forget" the blinded value of a bid since
it is not recoverable from the seed and blockchain data. This new rpc
call queries the txdb for the preimage (BlindValue) for a hash
(blind). It will be useful to inform the user that the bid value
is missing before the user tries to sendreveal and ends up with
a "Blind value not found" error. Wallet applications can test this
rpc call and inform the user they need to run `importnonce`.