wallet: import name
This enables a user to import a name into their wallet in a similar
way to importing an address or public key. A rescan is still required
to recover any auction activity and in particular the OPEN for a name
MUST be rescanned or the wallet will not record a nameState or track
any bids.

The WalletDB global name->wid map is used as a "watch" marker to
determine if a name's OPEN should be added to the txdb.