wallet: always save and recover critical namestate transitions
Modifies `txdb` during TX confirm and unconfirm:
If a TX updates namestate in a critical way, save it. Even if it
does not affect wallet balance or TX history. During rollback, revert,
reorg, or rescan -- look for the name undo data even if the TX
is not in wallet history, and apply.