test: do not pass entire MTX to http wallet client argument
The reason is that the MTX object will be converted to a generic
JSON object using MTX.getJSON() and then stringified. This process is
implemented in brq.RequestOptions.fromOptions(). The MTX object
in these tests had properties such as `fee: 0` and `rate: 0` which
get interpreted by the wallet node as options in mtx.fund(), resulting
in unexpected funding by the wallet.